// src/features/food/types.ts
export type FoodEntry = {
  id: string;
  date: string;
  concept: string;
  calories: number;
  proteinGrams: number;
  carbsGrams: number;
  createdAt: string;
  updatedAt: string;
};

export type FoodTemplate = {
  id: string;
  name: string;
  calories: number;
  proteinGrams: number;
  carbsGrams: number;
  createdAt: string;
  updatedAt: string;
};